requirements.txt
a2a-cli==0.2
a2a-json-rpc==0.3.0
a2a-sdk 
ag2==0.9.7
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
Authlib==1.6.5
autodoc_pydantic==2.2.0
autogen==0.9.7
autogen-agentchat==0.7.1
autogen-core==0.7.1
autogen-ext==0.7.1
Automat==25.4.16
azure-core==1.35.0
azure-identity==1.19.0
azure-storage-blob==12.26.0
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
botocore==1.36.6
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
constantly==23.10.4
create-mcp-server==1.0.5
cryptography==44.0.1
cssselect==1.3.0
cyclopts==4.1.0
dataclasses==0.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.21.2
dotenv==0.9.9
duckduckgo_search==8.1.1
ecs-logging==2.2.0
elastic-apm==6.24.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
fastapi-login==1.10.3
fastmcp==2.10.6
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
hyperlink==21.0.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
incremental==24.7.2
iniconfig==2.1.0
isodate==0.7.2
itemadapter==0.12.2
itemloaders==1.3.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json-schema-for-humans==1.3.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.75
langchain-mcp-adapters==0.1.9
langchain-ollama==0.3.6
langchain-openai==0.3.32
langchain-text-splitters==0.3.9
langgraph==0.5.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.8
lark==1.2.2
ldap3==2.9.1
limits==5.5.0
lxml==6.0.0
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.12.3
mdurl==0.1.2
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.48.1
neo4j==5.27.0
networkx==3.5
nltk==3.9.1
numpy==2.2.5
ollama==0.5.1
openai==1.61.1
openapi-pydantic==0.5.1
openpyxl==3.1.5
opentelemetry-api==1.36.0
ordered-set==4.1.0
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
parsel==1.10.0
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
plotly==6.2.0
pluggy==1.6.0
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.3.2
Protego==0.5.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
psycopg==3.2.9
psycopg-binary==3.2.9
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.19.2
PyJWT==2.10.1
pyodbc==5.2.0
pyOpenSSL==25.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.11.0
pytesseract==0.3.13
pytest==8.3.4
pytest-env==1.1.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.2
queuelib==1.8.0
RapidFuzz==3.14.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.3
requests==2.32.4
requests-aws4auth==1.3.1
requests-file==2.1.0
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==14.1.0
rich-rst==1.3.2
roman-numerals-py==3.1.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
Scrapy==2.13.3
sentence-transformers==5.0.0
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.2.3
sphinx-markdown-builder==0.6.8
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
sqlglot==26.23.0
sqlglotrs==0.5.0
sse-starlette==3.0.2
starlette==0.45.3
streamlit==1.47.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tldextract==5.3.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.54.0
Twisted==25.5.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.8.0
uvicorn==0.35.0
validators==0.22.0
w3lib==2.3.1
watchdog==6.0.0
wcwidth==0.2.13
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zope.interface==8.0
zstandard==0.23.0